Move enum declaration to the ty module.
