+++ sway-core/src/control_flow_analysis/dead_code_analysis.rs
@@ -4,7 +4,7 @@ use crate::{
     language::{parsed::TreeType, ty, CallPath, Visibility},
     semantic_analysis::{
         ast_node::{
-            TyAbiDeclaration, TyCodeBlock, TyConstantDeclaration, TyEnumDeclaration,
+            TyAbiDeclaration, TyCodeBlock, TyConstantDeclaration,
             TyFunctionDeclaration, TyStructDeclaration, TyStructExpressionField,
             TyTraitDeclaration, TyVariableDeclaration, VariableMutability,
         },
@@ -512,7 +512,7 @@ fn connect_abi_declaration(
 /// variant. When a variant is constructed, we can point an edge at that variant. This way,
 /// we can see clearly, and thusly warn, when individual variants are not ever constructed.
 fn connect_enum_declaration(
-    enum_decl: &TyEnumDeclaration,
+    enum_decl: &ty::TyEnumDeclaration,
     graph: &mut ControlFlowGraph,
     entry_node: NodeIndex,
 ) {
@@ -1186,7 +1186,7 @@ fn connect_code_block(
 }
 
 fn connect_enum_instantiation(
-    enum_decl: &TyEnumDeclaration,
+    enum_decl: &ty::TyEnumDeclaration,
     contents: &Option<Box<ty::TyExpression>>,
     variant_name: &Ident,
     graph: &mut ControlFlowGraph,
