+++ sway-core/src/language/ty/declaration/enum.rs
@@ -1,0 +1,111 @@
+
+use std::hash::{Hasher, Hash};
+
+use sway_types::{Ident, Span, Spanned};
+
+use crate::{
+    type_system::*, AttributesMap, language::Visibility,
+};
+
+#[derive(Clone, Debug, Eq)]
+pub struct TyEnumDeclaration {
+    pub name: Ident,
+    pub type_parameters: Vec<TypeParameter>,
+    pub attributes: AttributesMap,
+    pub variants: Vec<TyEnumVariant>,
+    pub(crate) span: Span,
+    pub visibility: Visibility,
+}
+
+// NOTE: Hash and PartialEq must uphold the invariant:
+// k1 == k2 -> hash(k1) == hash(k2)
+// https://doc.rust-lang.org/std/collections/struct.HashMap.html
+impl PartialEq for TyEnumDeclaration {
+    fn eq(&self, other: &Self) -> bool {
+        self.name == other.name
+            && self.type_parameters == other.type_parameters
+            && self.variants == other.variants
+            && self.visibility == other.visibility
+    }
+}
+
+impl CopyTypes for TyEnumDeclaration {
+    fn copy_types(&mut self, type_mapping: &TypeMapping) {
+        self.variants
+            .iter_mut()
+            .for_each(|x| x.copy_types(type_mapping));
+        self.type_parameters
+            .iter_mut()
+            .for_each(|x| x.copy_types(type_mapping));
+    }
+}
+
+impl CreateTypeId for TyEnumDeclaration {
+    fn create_type_id(&self) -> TypeId {
+        insert_type(TypeInfo::Enum {
+            name: self.name.clone(),
+            variant_types: self.variants.clone(),
+            type_parameters: self.type_parameters.clone(),
+        })
+    }
+}
+
+impl Spanned for TyEnumDeclaration {
+    fn span(&self) -> Span {
+        self.span.clone()
+    }
+}
+
+impl MonomorphizeHelper for TyEnumDeclaration {
+    fn type_parameters(&self) -> &[TypeParameter] {
+        &self.type_parameters
+    }
+
+    fn name(&self) -> &Ident {
+        &self.name
+    }
+}
+
+#[derive(Debug, Clone, Eq)]
+pub struct TyEnumVariant {
+    pub name: Ident,
+    pub type_id: TypeId,
+    pub initial_type_id: TypeId,
+    pub type_span: Span,
+    pub(crate) tag: usize,
+    pub(crate) span: Span,
+}
+
+// NOTE: Hash and PartialEq must uphold the invariant:
+// k1 == k2 -> hash(k1) == hash(k2)
+// https://doc.rust-lang.org/std/collections/struct.HashMap.html
+impl Hash for TyEnumVariant {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        self.name.hash(state);
+        look_up_type_id(self.type_id).hash(state);
+        self.tag.hash(state);
+    }
+}
+
+// NOTE: Hash and PartialEq must uphold the invariant:
+// k1 == k2 -> hash(k1) == hash(k2)
+// https://doc.rust-lang.org/std/collections/struct.HashMap.html
+impl PartialEq for TyEnumVariant {
+    fn eq(&self, other: &Self) -> bool {
+        self.name == other.name
+            && look_up_type_id(self.type_id) == look_up_type_id(other.type_id)
+            && self.tag == other.tag
+    }
+}
+
+impl CopyTypes for TyEnumVariant {
+    fn copy_types(&mut self, type_mapping: &TypeMapping) {
+        self.type_id.copy_types(type_mapping);
+    }
+}
+
+impl ReplaceSelfType for TyEnumVariant {
+    fn replace_self_type(&mut self, self_type: TypeId) {
+        self.type_id.replace_self_type(self_type);
+    }
+}
