+++ sway-core/src/semantic_analysis/ast_node/declaration/enum.rs
@@ -1,77 +1,13 @@
 use crate::{
     error::*,
-    language::{parsed::*, Visibility},
+    language::{parsed::*, ty},
     semantic_analysis::*,
-    type_system::{
-        insert_type, look_up_type_id, CopyTypes, CreateTypeId, EnforceTypeArguments,
-        MonomorphizeHelper, ReplaceSelfType, TypeId, TypeMapping, TypeParameter,
-    },
-    AttributesMap, TypeInfo,
+    type_system::*,
 };
-use std::hash::{Hash, Hasher};
 use sway_error::error::CompileError;
-use sway_types::{Ident, Span, Spanned};
+use sway_types::{Ident};
 
-#[derive(Clone, Debug, Eq)]
-pub struct TyEnumDeclaration {
-    pub name: Ident,
-    pub type_parameters: Vec<TypeParameter>,
-    pub attributes: AttributesMap,
-    pub variants: Vec<TyEnumVariant>,
-    pub(crate) span: Span,
-    pub visibility: Visibility,
-}
-
-// NOTE: Hash and PartialEq must uphold the invariant:
-// k1 == k2 -> hash(k1) == hash(k2)
-// https://doc.rust-lang.org/std/collections/struct.HashMap.html
-impl PartialEq for TyEnumDeclaration {
-    fn eq(&self, other: &Self) -> bool {
-        self.name == other.name
-            && self.type_parameters == other.type_parameters
-            && self.variants == other.variants
-            && self.visibility == other.visibility
-    }
-}
-
-impl CopyTypes for TyEnumDeclaration {
-    fn copy_types(&mut self, type_mapping: &TypeMapping) {
-        self.variants
-            .iter_mut()
-            .for_each(|x| x.copy_types(type_mapping));
-        self.type_parameters
-            .iter_mut()
-            .for_each(|x| x.copy_types(type_mapping));
-    }
-}
-
-impl CreateTypeId for TyEnumDeclaration {
-    fn create_type_id(&self) -> TypeId {
-        insert_type(TypeInfo::Enum {
-            name: self.name.clone(),
-            variant_types: self.variants.clone(),
-            type_parameters: self.type_parameters.clone(),
-        })
-    }
-}
-
-impl Spanned for TyEnumDeclaration {
-    fn span(&self) -> Span {
-        self.span.clone()
-    }
-}
-
-impl MonomorphizeHelper for TyEnumDeclaration {
-    fn type_parameters(&self) -> &[TypeParameter] {
-        &self.type_parameters
-    }
-
-    fn name(&self) -> &Ident {
-        &self.name
-    }
-}
-
-impl TyEnumDeclaration {
+impl ty::TyEnumDeclaration {
     pub fn type_check(ctx: TypeCheckContext, decl: EnumDeclaration) -> CompileResult<Self> {
         let mut errors = vec![];
         let mut warnings = vec![];
@@ -106,7 +42,7 @@ impl TyEnumDeclaration {
         let mut variants_buf = vec![];
         for variant in variants {
             variants_buf.push(check!(
-                TyEnumVariant::type_check(ctx.by_ref(), variant.clone()),
+                ty::TyEnumVariant::type_check(ctx.by_ref(), variant.clone()),
                 continue,
                 warnings,
                 errors
@@ -114,7 +50,7 @@ impl TyEnumDeclaration {
         }
 
         // create the enum decl
-        let decl = TyEnumDeclaration {
+        let decl = ty::TyEnumDeclaration {
             name,
             type_parameters: new_type_parameters,
             variants: variants_buf,
@@ -128,7 +64,7 @@ impl TyEnumDeclaration {
     pub(crate) fn expect_variant_from_name(
         &self,
         variant_name: &Ident,
-    ) -> CompileResult<&TyEnumVariant> {
+    ) -> CompileResult<&ty::TyEnumVariant> {
         let warnings = vec![];
         let mut errors = vec![];
         match self
@@ -149,51 +85,7 @@ impl TyEnumDeclaration {
     }
 }
 
-#[derive(Debug, Clone, Eq)]
-pub struct TyEnumVariant {
-    pub name: Ident,
-    pub type_id: TypeId,
-    pub initial_type_id: TypeId,
-    pub type_span: Span,
-    pub(crate) tag: usize,
-    pub(crate) span: Span,
-}
-
-// NOTE: Hash and PartialEq must uphold the invariant:
-// k1 == k2 -> hash(k1) == hash(k2)
-// https://doc.rust-lang.org/std/collections/struct.HashMap.html
-impl Hash for TyEnumVariant {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        self.name.hash(state);
-        look_up_type_id(self.type_id).hash(state);
-        self.tag.hash(state);
-    }
-}
-
-// NOTE: Hash and PartialEq must uphold the invariant:
-// k1 == k2 -> hash(k1) == hash(k2)
-// https://doc.rust-lang.org/std/collections/struct.HashMap.html
-impl PartialEq for TyEnumVariant {
-    fn eq(&self, other: &Self) -> bool {
-        self.name == other.name
-            && look_up_type_id(self.type_id) == look_up_type_id(other.type_id)
-            && self.tag == other.tag
-    }
-}
-
-impl CopyTypes for TyEnumVariant {
-    fn copy_types(&mut self, type_mapping: &TypeMapping) {
-        self.type_id.copy_types(type_mapping);
-    }
-}
-
-impl ReplaceSelfType for TyEnumVariant {
-    fn replace_self_type(&mut self, self_type: TypeId) {
-        self.type_id.replace_self_type(self_type);
-    }
-}
-
-impl TyEnumVariant {
+impl ty::TyEnumVariant {
     pub(crate) fn type_check(
         mut ctx: TypeCheckContext,
         variant: EnumVariant,
@@ -213,7 +105,7 @@ impl TyEnumVariant {
             errors,
         );
         ok(
-            TyEnumVariant {
+            ty::TyEnumVariant {
                 name: variant.name.clone(),
                 type_id: enum_variant_type,
                 initial_type_id,
