+++ sway-core/src/semantic_analysis/ast_node/expression/typed_expression/enum_instantiation.rs
@@ -13,7 +13,7 @@ use sway_types::{Ident, Spanned};
 #[allow(clippy::too_many_arguments)]
 pub(crate) fn instantiate_enum(
     ctx: TypeCheckContext,
-    enum_decl: TyEnumDeclaration,
+    enum_decl: ty::TyEnumDeclaration,
     enum_name: Ident,
     enum_variant_name: Ident,
     args: Vec<Expression>,
